Separate logic and export in generation mod.
