+++ generation/shared/simple_picker.lua
@@ -1,21 +1,10 @@
----@meta
-local generation = {}
-if false then _G.generation = generation end
+local simplePicker = {}
 
-generation.Generator = require("shared.Generator")
-generation.Picker = require("shared.Picker")
-
----@deprecated use generation.Generator instead.
-generation.LegacyGenerator = require("shared.LegacyGenerator")
-
----Randomly picks an item from the list.
----
----If you don't need weighted pick, consider using `table.pick_random` instead.
 ---@generic T
 ---@param itemsAndWeights {[1]:T,[2]:number}[] List of items and its weights.
 ---@param rng love.RandomGenerator? Random number generator to use.
 ---@return T
-function generation.pickWeighted(itemsAndWeights, rng)
+function simplePicker.pickWeighted(itemsAndWeights, rng)
     local weightSum = 0
 
     for _, itemAndWeight in ipairs(itemsAndWeights) do
@@ -42,15 +31,13 @@ function generation.pickWeighted(itemsAndWeights, rng)
     return nil
 end
 
----Randomly picks an item from the list.
----
----If all weights are in equal size pick, consider using `table.pick_random` instead.
+
 ---@generic T
 ---@param items T[] List of items.
 ---@param weights number[] List of item weights.
 ---@param rng love.RandomGenerator? Random number generator to use.
 ---@return T
-function generation.pickWeightedPlanar(items, weights, rng)
+function simplePicker.pickWeightedPlanar(items, weights, rng)
     local weightSum = 0
 
     for _, weight in ipairs(weights) do
@@ -77,6 +64,4 @@ function generation.pickWeightedPlanar(items, weights, rng)
     return nil
 end
 
-umg.expose("generation", generation)
-
-return generation
+return simplePicker
